pyproject.toml
[project]
name = "cog-sandbox-mcp"
version = "0.3.1"
description = "Cog OS filesystem-coding-agent sandbox exposed via the Model Context Protocol"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.25,<2",
]

[project.scripts]
cog-sandbox-mcp = "cog_sandbox_mcp.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
]
evals = [
    "httpx>=0.28",
    "pyyaml>=6",
    "watchdog>=6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cog_sandbox_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
